Clown fires more projectiles
Fixed the behavior of the purple  tomato to roll, the white tomato floats around, the orange tomato bounces across the ground, the mech can shoot a tomato sauce projectile from its nose ,and it can dash towards the player.
Need to make the tomato sauce fire in a stream.
Changed the tomato sauce sprite and moved some scripts into the script folder.