Boss Room and Clown Mech Animations
A Boss Room has been made but it seems to suffer from collision issues and Red Tomatos occasionally clip through the floor and the camera shows the outside of the room.

The Clown Mech now shows animations when it is selecting an attack and when an attack has been selected. Still need to add a firing animation and add two more projectiles.

Need the mech to pass through platforms.